+++ src/renderer/stores/history.js
