+++ src/renderer/stores/index.js
