+++ src/renderer/stores/settings.js
