+++ src/renderer/stores/utils.js
